requirements.txt
torch
torchsde
torchvision
einops
transformers>=4.25.1
safetensors>=0.3.0
aiohttp
pyyaml
Pillow
scipy
tqdm
psutil
rembg
numba
python-dotenv
runpod
websocket-client
supabase